# public/api.yaml
openapi: 3.0.0
info:
  title: 'Serveur API Documentation'
  description: 'Serveur API Documentation.'
  contact:
    name: 'Elsa Logier & Arnaud Fievet & Jules Bobeuf & Thomas Santoro'
    email: thomas_santoro@ens.univ-artois.fr
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  -
    url: /api
    description: 'OpenApi host'
  -
    url: 'http://127.0.0.1:8000/api'
    description: 'Ludotech API Server'
paths:
  /api/Register:
    post:
      tags:
        - Adherent
      summary: 'User Register'
      description: 'Register User Here'
      operationId: authRegister
      requestBody:
        content:
          application/json:
            schema: {  }
      responses:
        '200':
          description: 'Adherent created successfully'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Adherent'
  /api/login:
    post:
      tags:
        - Adherent
      summary: 'User Login'
      description: 'Login User Here'
      operationId: authLogin
      requestBody:
        content:
          application/json:
            schema: {  }
      responses:
        '200':
          description: 'Adherent logged successfully'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Adherent'
        '401':
          description: Unauthorized
...
  /api/info:
    get:
      tags:
        - Adherent
      summary: 'User info'
      description: 'User info here'
      operationId: Info
      responses:
        '200':
          description: 'Successfully logged out'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Achat'
        '403':
          description: Unauthorized
  /api/Edit:
    post:
      tags:
        - Adherent
      summary: 'User Edit'
      description: 'Edit User Here'
      operationId: authEdit
      requestBody:
        content:
          application/json:
            schema: {  }
      responses:
        '200':
          description: 'Adherent updated successfully'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Adherent'
  /api/Avatar:
    post:
      tags:
        - Adherent
      summary: 'User Avatar'
      description: 'Avatar User Here'
      operationId: authAvatar
      requestBody: {  }
      responses:
        '200':
          description: 'Adherent avatar updated successfully'
        '422':
          description: 'Adherent avatar updated successfully'
...
components:
  schemas:
    Adherent:
      required:
        - id
        - name
        - email
        - password
        - nom
        - prenom
        - pseudo
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        password:
          type: string
        nom:
          type: string
        prenom:
          type: string
        pseudo:
          type: string
      type: object
...
    Achat:
      required:
        - id
        - date_achat
        - lieu_achat
        - prix
      properties:
        id:
          type: integer
        date_achat:
          type: string
        lieu_achat:
          type: string
        prix:
          type: integer
      type: object
...